[SI-60] Bundle all external resources with Skylight [SI-62] Support for lightbox with common image formats [SI-63] Support for JWPlayer with common a/v formats
A lot of images and javascripts are moved with this refactor. A lot of custom themes will break as a result of this! Images are now in /assets/ instead of in default (and each other..) theme, as are all plugins like tinyMCE, jquery, etc.. (and their associated images). CSS remains in default/demo theme.

Lightbox now works out of box with additional config lines. JWPlayer is bundled but needs integration.